+++ projection/winui3/winui3-Microsoft.UI.Content/winui3/microsoft/ui/content/__init__.py
@@ -10,6 +10,7 @@ __all__ = [
     "ContentCoordinateRoundingMode",
     "ContentLayoutDirection",
     "ContentSizePolicy",
+    "PopupAnchor",
     "ChildSiteLink",
     "ContentCoordinateConverter",
     "ContentDeferral",