+++ projection/winui3/winui3-Microsoft.UI.Xaml.Automation.Peers/winui3/microsoft/ui/xaml/automation/peers/__init__.py
@@ -108,6 +108,7 @@ __all__ = [
     "SemanticZoomAutomationPeer",
     "SliderAutomationPeer",
     "SplitButtonAutomationPeer",
+    "SplitMenuFlyoutItemAutomationPeer",
     "TabViewAutomationPeer",
     "TabViewItemAutomationPeer",
     "TeachingTipAutomationPeer",