+++ projection/winui3/winui3-Microsoft.UI.Xaml.Markup/winui3/microsoft/ui/xaml/markup/__init__.py
@@ -20,6 +20,7 @@ __all__ = [
     "IRootObjectProvider",
     "IUriContext",
     "IXamlBindScopeDiagnostics",
+    "IXamlCondition",
     "IXamlMember",
     "IXamlMetadataProvider",
     "IXamlType",