+++ projection/winui3/winui3-Microsoft.Windows.AI.Imaging/winui3/microsoft/windows/ai/imaging/__init__.py
@@ -14,6 +14,7 @@ __all__ = [
     "ImageDescriptionResult",
     "ImageObjectExtractor",
     "ImageObjectExtractorHint",
+    "ImageObjectRemover",
     "ImageScaler",
     "RecognizedLine",
     "RecognizedText",