+++ projection/winui3/winui3-Microsoft.Windows.AI/winui3/microsoft/windows/ai/__init__.py
@@ -6,8 +6,10 @@ import winrt.system
 winrt.runtime._internals.load_projection(__spec__)
 
 __all__ = [
+    "AICapabilityCategory",
     "AIFeatureReadyResultState",
     "AIFeatureReadyState",
+    "AICapabilities",
     "AIFeatureReadyResult",
 ]