+++ projection/winui3/winui3-Microsoft.Windows.Widgets.Feeds.Providers/winui3/microsoft/windows/widgets/feeds/providers/__init__.py
@@ -23,6 +23,7 @@ __all__ = [
     "IFeedAnnouncementInvokedTarget",
     "IFeedManager",
     "IFeedManager2",
+    "IFeedManager3",
     "IFeedProvider",
     "IFeedProviderAnalytics",
     "IFeedProviderErrors",